refactor Api Service related to #26

+++ client/src/sevices/ApiService.ts
@@ -1,21 +1,23 @@
-import axios, { AxiosRequestConfig, AxiosResponse } from 'axios';
+import axios, { AxiosRequestConfig, AxiosResponse, AxiosInstance } from 'axios';
 
 import JwtService from './JwtService';
 
 export default class ApiServices {
-    private static axios = axios;
+    private static axios: AxiosInstance | null;
 
     public static init():void {
-      this.axios.defaults.baseURL = process.env.BASE_URL;
-    }
-
-    public static setHeader(): void {
-      this.axios.defaults.headers.common.Authorization = `Bearer ${JwtService.getToken()}`;
-      this.axios.defaults.headers.common.Accept = 'application/json';
-      this.axios.defaults.headers.common['Content-Type'] = 'application/json';
+      this.axios = axios.create({
+        baseURL: process.env.BASE_URL,
+        headers: {
+          Authorization: `Bearer ${JwtService.getToken()}`,
+          accept: 'application/json',
+          'Content-Type': 'application/json',
+        },
+      });
     }
 
     public static get(endPoint:string, config?:AxiosRequestConfig):Promise<AxiosResponse> {
+      if (!this.axios) throw new Error('Axios instance not initialized');
       return this.axios.get(endPoint, config);
     }
 
@@ -24,6 +26,7 @@ export default class ApiServices {
       body:any,
       config?:AxiosRequestConfig,
     ):Promise<AxiosResponse> {
+      if (!this.axios) throw new Error('Axios instance not initialized');
       return this.axios.post(endPoint, body, config);
     }
 
@@ -32,6 +35,7 @@ export default class ApiServices {
       body:any,
       config?:AxiosRequestConfig,
     ):Promise<AxiosResponse> {
+      if (!this.axios) throw new Error('Axios instance not initialized');
       return this.axios.put(endPoint, body, config);
     }
 
@@ -39,6 +43,7 @@ export default class ApiServices {
       endPoint:string,
       config?:AxiosRequestConfig,
     ):Promise<AxiosResponse> {
+      if (!this.axios) throw new Error('Axios instance not initialized');
       return this.axios.delete(endPoint, config);
     }
 }
